+++ packages/ui-search/CHANGELOG.md
@@ -1,5 +1,66 @@
 # @sharelyai/ui-search
 
+## 0.0.18
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @sharelyai/ui-shared@0.0.18
+  - @sharelyai/services@0.0.21
+
+## 0.0.17
+
+### Patch Changes
+
+- Bump all packages to sync versions
+- Updated dependencies
+  - @sharelyai/services@0.0.20
+  - @sharelyai/ui-shared@0.0.17
+
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @sharelyai/services@0.0.19
+  - @sharelyai/ui-shared@0.0.16
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @sharelyai/services@0.0.18
+  - @sharelyai/ui-shared@0.0.15
+
+## 0.0.14
+
+### Patch Changes
+
+- update sources
+- Updated dependencies
+  - @sharelyai/ui-shared@0.0.14
+  - @sharelyai/services@0.0.17
+
+## 0.0.13
+
+### Patch Changes
+
+- update validations
+- Updated dependencies
+  - @sharelyai/ui-shared@0.0.13
+  - @sharelyai/services@0.0.15
+
+## 0.0.12
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @sharelyai/ui-shared@0.0.12
+  - @sharelyai/services@0.0.14
+
 ## 0.0.10
 
 ### Patch Changes